requirements.txt
openai
python-dotenv
requests
pandas
scikit-learn
kagglehub
pillow
transformers
torch
